proyect finished

+++ RESTAURANT/next-restaurant/pages/api/products/[id].js
@@ -5,8 +5,12 @@ export default async function handler(req, res) {
   const { 
     method, 
     query:{id}, 
+    cookies,
   } = req;
+  const token = cookies.token;
+
   await dbConnect();
+
   if(method === "GET"){
     try{
       const product = await Product.findById(id);
@@ -16,6 +20,8 @@ export default async function handler(req, res) {
     }
   }
   if(method === "PUT"){
+    if(!token || token !== process.env.TOKEN)
+      return res.status(401).json("Not authenticated");
     try{
       const product = await Product.findByIdAndUpdate(id,req.body, {
         new: true,
@@ -27,6 +33,8 @@ export default async function handler(req, res) {
     }
   }
   if(method === "DELETE"){
+    if(!token || token !== process.env.TOKEN)
+      return res.status(401).json("Not authenticated");
     try{
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted!");
